+++ Sources/StytchCore/StytchClient/MagicLinks/Email/Email.swift
@@ -31,23 +31,3 @@ public extension StytchClient.MagicLinks {
     /// The interface for interacting with email magic links.
     var email: Email { .init(pathContext: pathContext) }
 }
-
-// private extension StytchClient.MagicLinks.Email {
-struct CodeChallengedParameters<T: Encodable>: Encodable {
-    private enum CodingKeys: String, CodingKey { case codeChallenge, codeChallengeMethod }
-
-    let codeChallenge: String
-    let codeChallengeMethod: String
-    let wrapped: T
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-
-        try wrapped.encode(to: encoder)
-
-        try container.encode(codeChallenge, forKey: .codeChallenge)
-        try container.encode(codeChallengeMethod, forKey: .codeChallengeMethod)
-    }
-}
-
-// }
