+++ Sources/StytchCore/StytchClient/MagicLinks/MagicLinks.swift
@@ -23,21 +23,3 @@ public extension StytchClient {
     /// The interface for interacting with magic-links products.
     static var magicLinks: MagicLinks { .init() }
 }
-
-// private extension StytchClient.MagicLinks {
-struct CodeVerifierParameters<T: Encodable>: Encodable {
-    private enum CodingKeys: String, CodingKey { case codeVerifier }
-
-    let codeVerifier: String
-    let wrapped: T
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-
-        try wrapped.encode(to: encoder)
-
-        try container.encode(codeVerifier, forKey: .codeVerifier)
-    }
-}
-
-// }
